// gtfs-processor/shared/gtfs-db-types/types.ts
export enum CongestionLevel {
  UNKNOWN_CONGESTION_LEVEL = "UNKNOWN_CONGESTION_LEVEL",
  RUNNING_SMOOTHLY = "RUNNING_SMOOTHLY",
  STOP_AND_GO = "STOP_AND_GO",
  CONGESTION = "CONGESTION",
  SEVERE_CONGESTION = "SEVERE_CONGESTION",
}

export enum Direction {
  OUTBOUND = "OUTBOUND",
  INBOUND = "INBOUND",
}

export enum LocationType {
  STOP_OR_PLATFORM = "STOP_OR_PLATFORM",
  STATION = "STATION",
  ENTRANCE_EXIT = "ENTRANCE_EXIT",
  GENERIC_NODE = "GENERIC_NODE",
  BOARDING_AREA = "BOARDING_AREA",
}

export enum OccupancyStatus {
  EMPTY = "EMPTY",
  MANY_SEATS_AVAILABLE = "MANY_SEATS_AVAILABLE",
  FEW_SEATS_AVAILABLE = "FEW_SEATS_AVAILABLE",
  STANDING_ROOM_ONLY = "STANDING_ROOM_ONLY",
  CRUSHED_STANDING_ROOM_ONLY = "CRUSHED_STANDING_ROOM_ONLY",
  FULL = "FULL",
  NOT_ACCEPTING_PASSENGERS = "NOT_ACCEPTING_PASSENGERS",
  NO_DATA_AVAILABLE = "NO_DATA_AVAILABLE",
  NOT_BOARDABLE = "NOT_BOARDABLE",
}

export enum PickupDropOff {
  REGULAR = "REGULAR",
  NO_PICKUP_OR_DROP_OFF = "NO_PICKUP_OR_DROP_OFF",
  PHONE_AGENCY = "PHONE_AGENCY",
  COORDINATE_WITH_DRIVER = "COORDINATE_WITH_DRIVER",
}

export enum RouteType {
  TRAM = "TRAM",
  SUBWAY = "SUBWAY",
  RAIL = "RAIL",
  BUS = "BUS",
  FERRY = "FERRY",
  CABLE_TRAM = "CABLE_TRAM",
  AERIAL_LIFT = "AERIAL_LIFT",
  FUNICULAR = "FUNICULAR",
  TROLLEYBUS = "TROLLEYBUS",
  MONORAIL = "MONORAIL",
}

export enum StopTimeUpdateScheduleRelationship {
  SCHEDULED = "SCHEDULED",
  SKIPPED = "SKIPPED",
  NO_DATA = "NO_DATA",
  UNSCHEDULED = "UNSCHEDULED",
}

export enum StrEnum {

}

export enum Timepoint {
  APPROXIMATE = "APPROXIMATE",
  EXACT = "EXACT",
}

export enum TripDescriptorScheduleRelationship {
  SCHEDULED = "SCHEDULED",
  ADDED = "ADDED",
  UNSCHEDULED = "UNSCHEDULED",
  CANCELED = "CANCELED",
  REPLACEMENT = "REPLACEMENT",
  DUPLICATED = "DUPLICATED",
  NEW = "NEW",
  DELETED = "DELETED",
}

export enum TripInstanceState {
  PRISTINE = "PRISTINE",
  DIRTY = "DIRTY",
  REMOVED = "REMOVED",
}

export enum VehicleStopStatus {
  INCOMING_AT = "INCOMING_AT",
  STOPPED_AT = "STOPPED_AT",
  IN_TRANSIT_TO = "IN_TRANSIT_TO",
}
// ...
export interface Route {
  _id?: null | string;
  revision_id?: null | string;
  agency_id: string;
  route_id: string;
  route_short_name?: null | string;
  route_long_name?: null | string;
  route_type: RouteType;
  route_color?: null | string;
  route_text_color?: null | string;
}
// ...
export interface StopTimeInstance {
  stop_id: string;
  stop_headsign?: null | string;
  pickup_type?: PickupDropOff | null;
  drop_off_type?: PickupDropOff | null;
  timepoint: Timepoint;
  shape_dist_traveled?: null | number;
  arrival_datetime?: Date | null;
  departure_datetime?: Date | null;
  predicted_arrival_datetime?: Date | null;
  predicted_departure_datetime?: Date | null;
  predicted_arrival_uncertainty?: null | number;
  predicted_departure_uncertainty?: null | number;
  schedule_relationship?: StopTimeUpdateScheduleRelationship | null;
}

export interface Trip {
  _id?: null | string;
  revision_id?: null | string;
  agency_id: string;
  trip_id: string;
  route_id: string;
  service_id: string;
  trip_headsign?: null | string;
  trip_short_name?: null | string;
  direction_id?: Direction | null;
  block_id?: null | string;
  shape_id?: null | string;
}
// ...
export interface VehiclePosition {
  _id?: null | string;
  revision_id?: null | string;
  agency_id: string;
  vehicle_id: string;
  timestamp?: Date;
  stop_id: null | string;
  current_stop_sequence: null | number;
  current_status: VehicleStopStatus | null;
  congestion_level: CongestionLevel | null;
  occupancy_status: OccupancyStatus | null;
  occupancy_percentage: null | number;
  latitude: null | number;
  longitude: null | number;
  bearing: null | number;
  odometer: null | number;
  speed: null | number;
  ingested_at?: Date;
  trip?: any | null;
}